refactor: move argtypes
