+++ .prettierignore
@@ -3,3 +3,4 @@ README.md
 docs
 .vscode
 .circleci
+dist/
