+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "cardano-hw-cli",
   "version": "1.12.0",
-  "commit": "c9bcf79f4b1053aa1af1cbaf3b602e3ed1879815",
+  "commit": "a557b2e6cda6a406302cfaa610bff6d10c965494",
   "description": "Cardano CLI tool for hardware wallets",
   "author": "Vacuumlabs",
   "homepage": "https://github.com/vacuumlabs/cardano-hw-cli#readme",
